feat: add crash dumps, file logs, and crash-watch
- Write daily process logs and capture crashes with stack traces
- Add crash-watch infra, compose, k8s, pulumi, and CI
- Include opcode in packet parse fatals
- Update runner local build and dist packaging
- Silence nodiscard warnings with std::ignore

+++ server/game/src/trade.cpp
@@ -282,7 +282,7 @@ uint8_t trade::add(uint8_t index)
     if (this->_items.contains(index))
         return this->_items.at(index);
 
-    auto order = this->_items.size();
+    auto order = static_cast<uint8_t>(this->_items.size());
     this->_items.insert({index, order});
     return order;
 }
